Revert "docs(): add sources summary page (#7480)"
This reverts commit c4d1d82a07afd73c44ccfcdd215c5b1dad35f8dc.
