+++ docs-website/docusaurus.config.js
@@ -11,9 +11,7 @@ module.exports = {
   favicon: "img/favicon.ico",
   organizationName: "datahub-project", // Usually your GitHub org/user name.
   projectName: "datahub", // Usually your repo name.
-  stylesheets: [
-    "https://fonts.googleapis.com/css2?family=Manrope:wght@400;600&display=swap",
-  ],
+  stylesheets: ["https://fonts.googleapis.com/css2?family=Manrope:wght@400;600&display=swap"],
   noIndex: isSaas,
   customFields: {
     isSaas: isSaas,
@@ -33,23 +31,14 @@ module.exports = {
       title: null,
       logo: {
         alt: "DataHub Logo",
-        src: `img/${
-          isSaas ? "acryl" : "datahub"
-        }-logo-color-light-horizontal.svg`,
-        srcDark: `img/${
-          isSaas ? "acryl" : "datahub"
-        }-logo-color-dark-horizontal.svg`,
+        src: `img/${isSaas ? "acryl" : "datahub"}-logo-color-light-horizontal.svg`,
+        srcDark: `img/${isSaas ? "acryl" : "datahub"}-logo-color-dark-horizontal.svg`,
       },
       items: [
         {
           to: "docs/",
           activeBasePath: "docs",
           label: "Docs",
-          position: "right",
-        },
-        {
-          href: "/integrations",
-          label: "Integrations",
           position: "right",
         },
         {
@@ -60,6 +49,11 @@ module.exports = {
         {
           href: "https://blog.datahubproject.io/",
           label: "Blog",
+          position: "right",
+        },
+        {
+          href: "https://feature-requests.datahubproject.io/",
+          label: "Feature Requests",
           position: "right",
         },
         {
@@ -194,24 +188,15 @@ module.exports = {
         blog: false,
         theme: {
           customCss: [
-            isSaas
-              ? require.resolve("./src/styles/acryl.scss")
-              : require.resolve("./src/styles/datahub.scss"),
+            isSaas ? require.resolve("./src/styles/acryl.scss") : require.resolve("./src/styles/datahub.scss"),
             require.resolve("./src/styles/global.scss"),
           ],
-        },
-        pages: {
-          path: "src/pages",
-          mdxPageComponent: "@theme/MDXPage",
         },
       },
     ],
   ],
   plugins: [
-    [
-      "@docusaurus/plugin-ideal-image",
-      { quality: 100, sizes: [320, 640, 1280, 1440, 1600] },
-    ],
+    ["@docusaurus/plugin-ideal-image", { quality: 100, sizes: [320, 640, 1280, 1440, 1600] }],
     "docusaurus-plugin-sass",
     [
       "docusaurus-graphql-plugin",
