+++ docs-website/src/pages/docs/_components/SearchBar/index.js
