+++ docs-website/src/pages/docs/index.js
@@ -20,17 +20,6 @@ import {
   HistoryOutlined,
 } from "@ant-design/icons";
 
-//quickLinkCards
-import {
-  ThunderboltTwoTone,
-  DeploymentUnitOutlined,
-  SyncOutlined,
-  CodeTwoTone,
-  QuestionCircleTwoTone,
-  SlidersTwoTone,
-  HeartTwoTone,
-} from "@ant-design/icons";
-
 const deploymentGuideContent = [
   {
     title: "Managed DataHub",
@@ -104,100 +93,17 @@ const ingestionGuideContent = [
 
 const featureGuideContent = [
   { title: "Domains", icon: <FolderTwoTone />, to: "docs/domains" },
-  {
-    title: "Glossary Terms",
-    icon: <BookTwoTone />,
-    to: "docs/glossary/business-glossary",
-  },
+  { title: "Glossary Terms", icon: <BookTwoTone />, to: "docs/glossary/business-glossary" },
   { title: "Tags", icon: <TagsTwoTone />, to: "docs/tags" },
-  {
-    title: "UI-Based Ingestion",
-    icon: <ApiTwoTone />,
-    to: "docs/ui-ingestion",
-  },
+  { title: "UI-Based Ingestion", icon: <ApiTwoTone />, to: "docs/ui-ingestion" },
   { title: "Search", icon: <SearchOutlined />, to: "docs/how/search" },
   // { title: "Browse", icon: <CompassTwoTone />, to: "/docs/quickstart" },
-  {
-    title: "Lineage Impact Analysis",
-    icon: <NodeExpandOutlined />,
-    to: "docs/act-on-metadata/impact-analysis",
-  },
-  {
-    title: "Metadata Tests",
-    icon: <CheckCircleTwoTone />,
-    to: "docs/tests/metadata-tests",
-  },
-  {
-    title: "Approval Flows",
-    icon: <SafetyCertificateTwoTone />,
-    to: "docs/managed-datahub/approval-workflows",
-  },
-  {
-    title: "Personal Access Tokens",
-    icon: <LockTwoTone />,
-    to: "docs/authentication/personal-access-tokens",
-  },
-  {
-    title: "Slack Notifications",
-    icon: <SlackOutlined />,
-    to: "docs/managed-datahub/saas-slack-setup",
-  },
-  {
-    title: "Schema History",
-    icon: <HistoryOutlined />,
-    to: "docs/schema-history",
-  },
-];
-
-const quickLinkContent = [
-  {
-    title: "Get Started",
-    icon: <ThunderboltTwoTone />,
-    description: "Details on how to get DataHub up and running",
-    to: "/docs/get-started-with-datahub",
-  },
-  {
-    title: "Ingest Metadata",
-    icon: <ApiTwoTone />,
-    description: "Details on how to get Metadata loaded into DataHub",
-    to: "/docs/metadata-ingestion",
-  },
-  {
-    title: "Enrich Metadata",
-    icon: <DeploymentUnitOutlined />,
-    description: "Improve the quality and coverage of Metadata",
-    to: "docs/enrich-metadata",
-  },
-  {
-    title: "Act on Metadata",
-    icon: <SyncOutlined />,
-    description: "Step-by-step guides for acting on Metadata Events",
-    to: "docs/act-on-metadata",
-  },
-  {
-    title: "Developer Guides",
-    icon: <CodeTwoTone />,
-    description: "Interact with DataHub programmatically ",
-    to: "/docs/cli",
-  },
-  {
-    title: "Feature Guides",
-    icon: <QuestionCircleTwoTone />,
-    description: "Step-by-step guides for making the most of DataHub",
-    to: "/docs/how/search",
-  },
-  {
-    title: "Deployment Guides",
-    icon: <SlidersTwoTone />,
-    description: "Step-by-step guides for deploying DataHub to production",
-    to: "/docs/deploy/aws",
-  },
-  {
-    title: "Join the Community",
-    icon: <HeartTwoTone />,
-    description: "Collaborate, learn, and grow with us",
-    to: "/docs/slack",
-  },
+  { title: "Lineage Impact Analysis", icon: <NodeExpandOutlined />, to: "docs/act-on-metadata/impact-analysis" },
+  { title: "Metadata Tests", icon: <CheckCircleTwoTone />, to: "docs/tests/metadata-tests" },
+  { title: "Approval Flows", icon: <SafetyCertificateTwoTone />, to: "docs/managed-datahub/approval-workflows" },
+  { title: "Personal Access Tokens", icon: <LockTwoTone />, to: "docs/authentication/personal-access-tokens" },
+  { title: "Slack Notifications", icon: <SlackOutlined />, to: "docs/managed-datahub/saas-slack-setup" },
+  { title: "Schema History", icon: <HistoryOutlined />, to: "docs/schema-history" },
 ];
 
 function Docs() {
@@ -214,27 +120,14 @@ function Docs() {
           <div className="hero__content">
             <div>
               <h1 className="hero__title">Documentation</h1>
-              <p className="hero__subtitle">
-                Guides and tutorials for everything DataHub.
-              </p>
+              <p className="hero__subtitle">Guides and tutorials for everything DataHub.</p>
               <SearchBar />
             </div>
           </div>
-          <QuickLinkCards quickLinkContent={quickLinkContent} />
-          <GuideList
-            title="Deployment Guides"
-            content={deploymentGuideContent}
-          />
-          <GuideList
-            title="Ingestion Guides"
-            content={ingestionGuideContent}
-            seeMoreLink={{ label: "See all 50+ sources", to: "/integrations" }}
-          />
-          <GuideList
-            title="Feature Guides"
-            content={featureGuideContent}
-            seeMoreLink={{ label: "See all guides", to: "/docs/how/search" }}
-          />
+          <QuickLinkCards />
+          <GuideList title="Deployment Guides" content={deploymentGuideContent} />
+          <GuideList title="Ingestion Guides" content={ingestionGuideContent} seeMoreLink={{ label: "See all 50+ sources", to: "/docs/lineage/airflow" }} />
+          <GuideList title="Feature Guides" content={featureGuideContent} seeMoreLink={{ label: "See all guides", to: "/docs/how/search" }} />
         </div>
       </header>
     </Layout>
