--- docs-website/src/pages/integrations.jsx
@@ -1,21 +1,0 @@
-import React from "react";
-
-import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
-import { FilterPage } from "./docs/_components/FilterPage";
-import { FastBackwardFilled } from "@ant-design/icons";
-const filterTagIndexes = require("../../filterTagIndexes.json");
-const metadata = filterTagIndexes.ingestionSources;
-
-function DataProviderComponent() {
-  const context = useDocusaurusContext();
-  const { siteConfig = {} } = context;
-
-  return FilterPage(
-    siteConfig,
-    metadata,
-    "DataHub Integrations",
-    "Services that integrate with DataHub"
-  );
-}
-
-export default DataProviderComponent;
