+++ docs-website/src/styles/global.scss
@@ -29,7 +29,7 @@
   /* Rule */
   --ifm-hr-border-width: 1px 0 0 0;
   --ifm-hr-border-color: #e3e3e3;
-  --ifm-hr-background-color: #e3e3e3;
+
   /* More Colors */
   --ifm-hero-background-color: var(--ifm-background-color);
   --ifm-background-surface-color: #fafafa;
